Add css styles and functionality to the "Settings" component

+++ client/src/components/Settings.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useState, useRef } from 'react';
 import Button from '@mui/material/Button';
 import IconButton from '@mui/material/IconButton';
 import SettingsIcon from '@mui/icons-material/Settings';
@@ -8,6 +8,9 @@ import Fade from '@mui/material/Fade';
 import Typography from '@mui/material/Typography';
 import Backdrop from '@mui/material/Backdrop';
 import TextField from '@mui/material/TextField';
+import { setMessage } from '../components/Snackbar';
+import ClearIcon from '@mui/icons-material/Clear';
+import { InputAdornment } from '@mui/material';
 import axios from 'axios';
 
 const style = {
@@ -26,29 +29,46 @@ const Settings = () => {
    const [open, setOpen] = useState(false);
    const handleOpen = () => setOpen(true);
    const handleClose = () => setOpen(false);
-
    const [newUsername, setNewUsername] = useState('');
    const [newPassword, setNewPassword] = useState('');
    const [deleteUsername, setDeleteUsername] = useState('');
+   const [isErrorUsername, setErrorUsername] = useState(false);
+   const [isErrorDeleteUser, setErrorDeleteUser] = useState(false);
+   const usernameInputRef = useRef(null);
+   const deleteInputRef = useRef(null);
 
    const createOrUpdateUser = async () => {
+      if (!newUsername.trim()) {
+         setMessage({ message: 'Field "username" cannot be empty!', color: 'error' });
+         setErrorUsername(true);
+         usernameInputRef.current.focus();
+         return;
+      }
+
       try {
          const response = await axios.post('/api/user/createOrUpdateUser', {
             username: newUsername,
             password: newPassword,
          });
          console.log(response.data.message);
-         console.log('User created');
+         setMessage({ message: 'User created or updated', color: 'success' });
       } catch (error) {
          console.error('Error creating or updating user:', error);
       }
    };
 
    const deleteUser = async (username) => {
+      if (!username.trim()) {
+         setMessage({ message: 'Field "delete user" cannot be empty!', color: 'error' });
+         setErrorDeleteUser(true);
+         deleteInputRef.current.focus();
+         return;
+      }
+
       try {
          const response = await axios.delete(`/api/user/deleteUser/${username}`);
          console.log(response.data.message);
-         console.log('User deleted');
+         setMessage({ message: 'User deleted', color: 'error' });
       } catch (error) {
          console.error('Error deleting user:', error);
       }
@@ -58,14 +78,24 @@ const Settings = () => {
       try {
          const response = await axios.get('/api/user/getAllUserLogins');
          const logins = response.data.logins;
-         console.log('User Logins:', logins);
+         console.log('A list of users:', logins);
+         setMessage({ message: `A list of users: ${logins.join(', ')}`, color: 'error' });
 
-         // Выводим сообщение в консоли браузера
          // alert('User Logins: ' + logins.join(', '));
       } catch (error) {
          console.error('Error getting user logins:', error);
       }
    };
+
+   const ClearButton = ({ value, onClick }) => (
+      <InputAdornment position="end">
+         {value && (
+            <IconButton className="clear-btn" onClick={onClick}>
+               <ClearIcon />
+            </IconButton>
+         )}
+      </InputAdornment>
+   );
 
    return (
       <>
@@ -90,41 +120,65 @@ const Settings = () => {
                <Box className="settings-modal-windows" sx={style}>
                   <div className="settings-input">
                      <Typography id="transition-modal-title" variant="h6" component="h2">
-                        Settings
+                        Settings:
                      </Typography>
 
                      <TextField
+                        className="input"
                         type="text"
-                        label="new username"
+                        label="username"
                         variant="standard"
                         autoComplete="off"
                         value={newUsername}
-                        onChange={(e) => setNewUsername(e.target.value)}
+                        onChange={(e) => {
+                           setNewUsername(e.target.value);
+                           setErrorUsername(false);
+                        }}
+                        onBlur={() => setErrorUsername(false)}
+                        error={isErrorUsername}
+                        inputRef={usernameInputRef}
+                        InputProps={{
+                           endAdornment: <ClearButton value={newUsername} onClick={() => setNewUsername('')} />,
+                        }}
                      />
 
                      <TextField
+                        className="input"
                         type="password"
-                        label="new password"
+                        label="password"
                         variant="standard"
                         autoComplete="off"
                         value={newPassword}
                         onChange={(e) => setNewPassword(e.target.value)}
+                        InputProps={{
+                           endAdornment: <ClearButton value={newPassword} onClick={() => setNewPassword('')} />,
+                        }}
                      />
 
-                     <Button className="btn-save" variant="outlined" onClick={createOrUpdateUser}>
+                     <Button className="btn-settings" variant="outlined" onClick={createOrUpdateUser}>
                         Create or update user
                      </Button>
 
                      <TextField
+                        className="input"
                         type="text"
                         label="delete user"
                         variant="standard"
                         autoComplete="off"
                         value={deleteUsername}
-                        onChange={(e) => setDeleteUsername(e.target.value)}
+                        onChange={(e) => {
+                           setDeleteUsername(e.target.value);
+                           setErrorDeleteUser(false);
+                        }}
+                        onBlur={() => setErrorDeleteUser(false)}
+                        error={isErrorDeleteUser}
+                        inputRef={deleteInputRef}
+                        InputProps={{
+                           endAdornment: <ClearButton value={deleteUsername} onClick={() => setDeleteUsername('')} />,
+                        }}
                      />
 
-                     <Button className="btn-save" variant="outlined" onClick={() => deleteUser(deleteUsername)}>
+                     <Button className="btn-settings" variant="outlined" onClick={() => deleteUser(deleteUsername)}>
                         Delete User
                      </Button>
 
@@ -147,11 +201,7 @@ const Settings = () => {
                         autoComplete="off"
                      />
 
-                     {/* <Button className="btn-save" variant="outlined">
-                        Save
-                     </Button> */}
-
-                     <Button className="btn-save" variant="outlined" onClick={() => getAllUserLogins()}>
+                     <Button className="btn-settings" variant="outlined" onClick={() => getAllUserLogins()}>
                         Get All User Logins
                      </Button>
                   </div>
